+++ app/pom.xml
@@ -181,11 +181,6 @@
             <artifactId>testcontainers-keycloak</artifactId>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>org.keycloak</groupId>
-            <artifactId>keycloak-admin-client</artifactId>
-            <scope>test</scope>
-        </dependency>
     </dependencies>
 
     <build>
