+++ app/src/test/java/io/apicurio/registry/AbstractResourceTestBase.java
@@ -20,11 +20,18 @@ import io.apicurio.registry.client.RegistryRestClient;
 import io.apicurio.registry.client.RegistryRestClientFactory;
 import io.apicurio.registry.types.ArtifactState;
 import io.apicurio.registry.types.ArtifactType;
+import io.apicurio.registry.util.ServiceInitializer;
 import io.apicurio.registry.utils.tests.TestUtils;
 import io.restassured.RestAssured;
 import io.restassured.response.ValidatableResponse;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.TestInstance.Lifecycle;
+
+import javax.enterprise.inject.Instance;
+import javax.inject.Inject;
+
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -33,15 +40,20 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
  * Abstract base class for all tests that test via the jax-rs layer.
  * @author eric.wittmann@gmail.com
  */
-public abstract class AbstractResourceTestBase extends RegistryTestBase {
+@TestInstance(Lifecycle.PER_CLASS)
+public abstract class AbstractResourceTestBase extends AbstractRegistryTestBase {
 
     protected static final String CT_JSON = "application/json";
     protected static final String CT_PROTO = "application/x-protobuf";
     protected static final String CT_YAML = "application/x-yaml";
     protected static final String CT_XML = "application/xml";
 
+    @Inject
+    Instance<ServiceInitializer> initializers;
+
     protected String registryUrl;
     protected RegistryRestClient client;
+
 
     @BeforeAll
     protected void beforeAll() throws Exception {
@@ -49,12 +61,17 @@ public abstract class AbstractResourceTestBase extends RegistryTestBase {
         client = RegistryRestClientFactory.create(registryUrl);
     }
 
-    @Override
     @BeforeEach
     protected void beforeEach() throws Exception {
-        super.beforeEach();
-        RestAssured.baseURI = registryUrl;
+        prepareServiceInitializers();
         deleteGlobalRules(0);
+    }
+
+    protected void prepareServiceInitializers() {
+        RestAssured.baseURI = registryUrl;
+
+        // run all initializers::beforeEach
+        initializers.stream().forEach(ServiceInitializer::beforeEach);
     }
 
     protected void deleteGlobalRules(int expectedDefaultRulesCount) throws Exception {
@@ -222,4 +239,3 @@ public abstract class AbstractResourceTestBase extends RegistryTestBase {
         response.body("state", equalTo(state.name()));
     }
 }
-
